Move logic to the sdk
